Move more module map code out of bindings and runtime
